Send chunks over the network
This commit is not (yet) perfect.

Basically, when we send chunks too fast, the client gets confused and some packets
are lost. So to solve this problem, we add a litte delay after sending each chunks.
I really tried hard to solve this issue, but without luck.

This commit also refactored the architecture of our sources so that the code
related with network communication is put into its own module.